+++ src/tourist/management/ui/layouts/GiaoDienDatVe.java
@@ -1,4 +1,4 @@
-package tourist.management.ui.components;
+package tourist.management.ui.layouts;
 
 import tourist.management.dao.ChuyenDiDAO;
 import tourist.management.dao.DatVeDAO;
