+++ src/tourist/management/ui/windows/GiaoDienChinh.java
@@ -1,4 +1,4 @@
-package tourist.management.ui.views;
+package tourist.management.ui.windows;
 
 import tourist.management.entity.NhanVien;
 import tourist.management.ui.layouts.*;
